Modify spriteFrame and Frame class.
